requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
Django==5.0.7
django-appconf==1.0.6
django-browser-reload==1.13.0
django-compressor==4.5.1
django-jazzmin==3.0.0
django-tailwind==3.8.0
et-xmlfile==1.1.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openpyxl==3.1.5
pillow==10.4.0
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.1
rcssmin==1.1.2
requests==2.32.3
rich==13.7.1
rjsmin==1.2.2
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
urllib3==2.2.2
whitenoise==6.7.0
